requirements.txt
neo4j>=5.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scrapy>=2.11.0
hyperbrowser>=0.3.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.0

